requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
cryptography==42.0.0
bcrypt==4.1.2
flask-limiter==3.5.0
psycopg2-binary==2.9.9
